NaturalEarth

Documentation for NaturalEarth.

NaturalEarth.bathymetry

NaturalEarth.naturalearth

NaturalEarth.naturalearth — Function

naturalearth(name::String; version = v"5.1.2")
naturalearth(name::String, scale::Int; version = v"5.1.2")

Load a NaturalEarth dataset as a GeoJSON.FeatureCollection object.

The name should not include the ne_ prefix, and if providing a scale should also not include a scale. No suffix should be added.

For example, to get the ne_110m_admin_0_countries dataset, you could use:

naturalearth("admin_0_countries", 110)
naturalearth("110m_admin_0_countries")

We aim to support all datasets listed in https://github.com/nvkelso/natural-earth-vector/tree/master/geojson.

Warning

This function downloads files from the Internet when not found locally.

NaturalEarth.bathymetry — Function

bathymetry(contour::Int = 2000)

Convenient acccess to ocean bathymetry datasets. Currently tested on: https://www.naturalearthdata.com/downloads/10m-physical-vectors/10m-bathymetry/

The function returns a MultiPolygon describing the bathymetry at a given depth contour. The following depths should be available: [10000, 9000, 8000, 7000, 6000, 5000, 4000, 3000, 2000, 1000, 200, 0]